changed the file quite a bit, not nearly done tho
a lot of work, not worth adding a description other than the receiving works, I believe

added a bunch of new data, still need to figure out how to analyze neighbors in a single array

will print to the file correctly given random number of stuff

finished some stuff, not nearly enough, but some

can now compare numbers correctly

not going to lie, this is janky

the zero process has an issue, idk why honestly so I kind of had a bit of a change

updated the gitignore. Essentially done

changed the values to actual black and white as opposed to other issues. about to screw the heck out of it

got the first process to work

has a reporting section